Re-implemented metadata
This closes #43
